server/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.109.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
# passlib is no longer used due to Python 3.12 compatibility issues

pyasn1==0.6.3
pycparser==3.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.2.2
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.15.0
uvicorn==0.27.0
